Hide `(Http|Request|Response)HeaderGetters` from the public API
