+++ core/src/main/java/com/linecorp/armeria/common/DefaultRequestHeaders.java
@@ -60,7 +60,7 @@ final class DefaultRequestHeaders extends DefaultHttpHeaders implements RequestH
 
     @Override
     public RequestHeadersBuilder toBuilder() {
-        return new RequestHeadersBuilder(this);
+        return new DefaultRequestHeadersBuilder(this);
     }
 
     @Override
