+++ core/src/main/java/com/linecorp/armeria/common/HttpHeadersBuilder.java
@@ -1,127 +1,22 @@
-/*
- * Copyright 2019 LINE Corporation
- *
- * LINE Corporation licenses this file to you under the Apache License,
- * version 2.0 (the "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at:
- *
- *   https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- */
 package com.linecorp.armeria.common;
 
-import static com.google.common.base.Preconditions.checkArgument;
-import static com.google.common.base.Preconditions.checkState;
-import static java.util.Objects.requireNonNull;
-
-import java.util.Collections;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map.Entry;
-import java.util.Set;
-import java.util.function.BiConsumer;
-import java.util.function.Consumer;
-import java.util.stream.Stream;
 
 import javax.annotation.Nullable;
-
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableSet;
-
-import io.netty.util.AsciiString;
 
 /**
  * Builds an {@link HttpHeaders}.
  *
- * @see RequestHeadersBuilder
- * @see ResponseHeadersBuilder
+ * @see HttpHeaders#builder()
+ * @see HttpHeaders#toBuilder()
  */
-public class HttpHeadersBuilder implements HttpHeaderGetters {
-
-    private int sizeHint = HttpHeadersBase.DEFAULT_SIZE_HINT;
-    @Nullable
-    private HttpHeadersBase delegate;
-    @Nullable
-    private HttpHeadersBase parent;
-
-    HttpHeadersBuilder() {}
-
-    HttpHeadersBuilder(HttpHeadersBase parent) {
-        assert parent instanceof HttpHeaders;
-        this.parent = parent;
-    }
-
+public interface HttpHeadersBuilder extends HttpHeaderGetters {
     /**
      * Returns a newly created {@link HttpHeaders} with the entries in this builder.
      * Note that any further modification of this builder is prohibited after this method is invoked.
      */
-    public HttpHeaders build() {
-        if (delegate != null) {
-            if (delegate.isEmpty()) {
-                return delegate.isEndOfStream() ? DefaultHttpHeaders.EMPTY_EOS : DefaultHttpHeaders.EMPTY;
-            } else {
-                return new DefaultHttpHeaders(promoteDelegate());
-            }
-        }
-
-        return parent != null ? (HttpHeaders) parent : DefaultHttpHeaders.EMPTY;
-    }
-
-    @Nullable
-    final HttpHeadersBase delegate() {
-        return delegate;
-    }
-
-    @Nullable
-    final HttpHeadersBase parent() {
-        return parent;
-    }
-
-    /**
-     * Makes the current {@link #delegate()} a new {@link #parent()} and clears the current {@link #delegate()}.
-     * Call this method when you create a new {@link HttpHeaders} derived from the {@link #delegate()},
-     * so that any further modifications to this builder do not break the immutability of {@link HttpHeaders}.
-     *
-     * @return the {@link #delegate()}
-     */
-    final HttpHeadersBase promoteDelegate() {
-        final HttpHeadersBase delegate = this.delegate;
-        assert delegate != null;
-        parent = delegate;
-        this.delegate = null;
-        return delegate;
-    }
-
-    @Nullable
-    final HttpHeadersBase getters() {
-        if (delegate != null) {
-            return delegate;
-        }
-
-        if (parent != null) {
-            return parent;
-        }
-
-        return null;
-    }
-
-    final HttpHeadersBase setters() {
-        if (delegate != null) {
-            return delegate;
-        }
-
-        if (parent != null) {
-            // Make a deep copy of the parent.
-            return delegate = new HttpHeadersBase(parent, false);
-        }
-
-        return delegate = new HttpHeadersBase(sizeHint);
-    }
+    HttpHeaders build();
 
     /**
      * Specifies the hint about the number of headers which may improve the memory efficiency and performance
@@ -131,242 +26,17 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @throws IllegalStateException if the hint was specified too late after the underlying data structure
      *                               has been fully initialized.
      */
-    public HttpHeadersBuilder sizeHint(int sizeHint) {
-        checkArgument(sizeHint >= 0, "sizeHint: %s (expected: >= 0)", sizeHint);
-        checkState(delegate == null, "sizeHint cannot be specified after a modification is made.");
-        this.sizeHint = sizeHint;
-        return this;
-    }
-
-    // Shortcuts
-
-    @Override
-    @Nullable
-    public final MediaType contentType() {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.contentType() : null;
-    }
-
-    /**
-     * Sets the {@code "content-type"} header.
-     */
-    public HttpHeadersBuilder contentType(MediaType contentType) {
-        setters().contentType(contentType);
-        return this;
-    }
-
-    // Getters
-
-    @Override
-    public final boolean isEndOfStream() {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.isEndOfStream() : false;
-    }
-
-    @Override
-    @Nullable
-    public final String get(CharSequence name) {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.get(name) : null;
-    }
-
-    @Override
-    public final String get(CharSequence name, String defaultValue) {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.get(name, defaultValue)
-                               : requireNonNull(defaultValue, "defaultValue");
-    }
-
-    @Override
-    public final List<String> getAll(CharSequence name) {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.getAll(name) : ImmutableList.of();
-    }
-
-    @Override
-    @Nullable
-    public final Integer getInt(CharSequence name) {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.getInt(name) : null;
-    }
-
-    @Override
-    public final int getInt(CharSequence name, int defaultValue) {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.getInt(name, defaultValue) : defaultValue;
-    }
-
-    @Override
-    @Nullable
-    public final Long getLong(CharSequence name) {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.getLong(name) : null;
-    }
-
-    @Override
-    public final long getLong(CharSequence name, long defaultValue) {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.getLong(name, defaultValue) : defaultValue;
-    }
-
-    @Override
-    @Nullable
-    public final Float getFloat(CharSequence name) {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.getFloat(name) : null;
-    }
-
-    @Override
-    public final float getFloat(CharSequence name, float defaultValue) {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.getFloat(name, defaultValue) : defaultValue;
-    }
-
-    @Override
-    @Nullable
-    public final Double getDouble(CharSequence name) {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.getDouble(name) : null;
-    }
-
-    @Override
-    public final double getDouble(CharSequence name, double defaultValue) {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.getDouble(name, defaultValue) : defaultValue;
-    }
-
-    @Override
-    @Nullable
-    public final Long getTimeMillis(CharSequence name) {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.getTimeMillis(name) : null;
-    }
-
-    @Override
-    public final long getTimeMillis(CharSequence name, long defaultValue) {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.getTimeMillis(name, defaultValue) : defaultValue;
-    }
-
-    @Override
-    public final boolean contains(CharSequence name) {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.contains(name) : false;
-    }
-
-    @Override
-    public final boolean contains(CharSequence name, String value) {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.contains(name, value) : false;
-    }
-
-    @Override
-    public final boolean containsObject(CharSequence name, Object value) {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.containsObject(name, value) : false;
-    }
-
-    @Override
-    public final boolean containsInt(CharSequence name, int value) {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.containsInt(name, value) : false;
-    }
-
-    @Override
-    public final boolean containsLong(CharSequence name, long value) {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.containsLong(name, value) : false;
-    }
-
-    @Override
-    public final boolean containsFloat(CharSequence name, float value) {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.containsFloat(name, value) : false;
-    }
-
-    @Override
-    public final boolean containsDouble(CharSequence name, double value) {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.containsDouble(name, value) : false;
-    }
-
-    @Override
-    public final boolean containsTimeMillis(CharSequence name, long value) {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.containsTimeMillis(name, value) : false;
-    }
-
-    @Override
-    public final int size() {
-        if (delegate != null) {
-            return delegate.size();
-        }
-        if (parent != null) {
-            return parent.size();
-        }
-        return 0;
-    }
-
-    @Override
-    public final boolean isEmpty() {
-        return size() == 0;
-    }
-
-    @Override
-    public final Set<AsciiString> names() {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.names() : ImmutableSet.of();
-    }
-
-    @Override
-    public final Iterator<Entry<AsciiString, String>> iterator() {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.iterator() : Collections.emptyIterator();
-    }
-
-    @Override
-    public final Iterator<String> valueIterator(CharSequence name) {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.valueIterator(name) : Collections.emptyIterator();
-    }
-
-    @Override
-    public final void forEach(BiConsumer<AsciiString, String> action) {
-        final HttpHeadersBase getters = getters();
-        if (getters != null) {
-            getters.forEach(action);
-        }
-    }
-
-    @Override
-    public final void forEachValue(CharSequence name, Consumer<String> action) {
-        final HttpHeadersBase getters = getters();
-        if (getters != null) {
-            getters.forEachValue(name, action);
-        }
-    }
-
-    @Override
-    public final Stream<Entry<AsciiString, String>> stream() {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.stream() : Stream.empty();
-    }
-
-    @Override
-    public final Stream<String> valueStream(CharSequence name) {
-        final HttpHeadersBase getters = getters();
-        return getters != null ? getters.valueStream(name) : Stream.empty();
-    }
-
-    // Setters
+    HttpHeadersBuilder sizeHint(int sizeHint);
 
     /**
      * Sets whether the headers will be the last frame in an HTTP/2 stream.
      */
-    public HttpHeadersBuilder endOfStream(boolean endOfStream) {
-        setters().endOfStream(endOfStream);
-        return this;
-    }
+    HttpHeadersBuilder endOfStream(boolean endOfStream);
+
+    /**
+     * Sets the {@code "content-type"} header.
+     */
+    HttpHeadersBuilder contentType(MediaType contentType);
 
     /**
      * Removes all the headers with the specified name and returns the header value which was added first.
@@ -375,9 +45,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @return the first header value or {@code null} if there is no such header
      */
     @Nullable
-    public final String getAndRemove(CharSequence name) {
-        return contains(name) ? setters().getAndRemove(name) : null;
-    }
+    String getAndRemove(CharSequence name);
 
     /**
      * Removes all the headers with the specified name and returns the header value which was added first.
@@ -386,10 +54,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param defaultValue the default value
      * @return the first header value or {@code defaultValue} if there is no such header
      */
-    public final String getAndRemove(CharSequence name, String defaultValue) {
-        return contains(name) ? setters().getAndRemove(name, defaultValue)
-                              : requireNonNull(defaultValue, "defaultValue");
-    }
+    String getAndRemove(CharSequence name, String defaultValue);
 
     /**
      * Removes all the headers with the specified name and returns the removed header values.
@@ -397,9 +62,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param name the name of the header to retrieve
      * @return a {@link List} of header values or an empty {@link List} if no values are found.
      */
-    public final List<String> getAllAndRemove(CharSequence name) {
-        return contains(name) ? setters().getAllAndRemove(name) : ImmutableList.of();
-    }
+    List<String> getAllAndRemove(CharSequence name);
 
     /**
      * Removes all the headers with the specified name and returns the header value which was added first.
@@ -409,9 +72,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      *         such value or it can't be converted to {@code int}.
      */
     @Nullable
-    public final Integer getIntAndRemove(CharSequence name) {
-        return contains(name) ? setters().getIntAndRemove(name) : null;
-    }
+    Integer getIntAndRemove(CharSequence name);
 
     /**
      * Removes all the headers with the specified name and returns the header value which was added first.
@@ -421,9 +82,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @return the {@code int} value of the first value in insertion order or {@code defaultValue} if there is
      *         no such value or it can't be converted to {@code int}.
      */
-    public final int getIntAndRemove(CharSequence name, int defaultValue) {
-        return contains(name) ? setters().getIntAndRemove(name, defaultValue) : defaultValue;
-    }
+    int getIntAndRemove(CharSequence name, int defaultValue);
 
     /**
      * Removes all the headers with the specified name and returns the header value which was added first.
@@ -433,9 +92,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      *         value or it can't be converted to {@code long}.
      */
     @Nullable
-    public final Long getLongAndRemove(CharSequence name) {
-        return contains(name) ? setters().getLongAndRemove(name) : null;
-    }
+    Long getLongAndRemove(CharSequence name);
 
     /**
      * Removes all the headers with the specified name and returns the header value which was added first.
@@ -445,9 +102,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @return the {@code long} value of the first value in insertion order or {@code defaultValue} if there is
      *         no such value or it can't be converted to {@code long}.
      */
-    public final long getLongAndRemove(CharSequence name, long defaultValue) {
-        return contains(name) ? setters().getLongAndRemove(name, defaultValue) : defaultValue;
-    }
+    long getLongAndRemove(CharSequence name, long defaultValue);
 
     /**
      * Removes all the headers with the specified name and returns the header value which was added first.
@@ -457,9 +112,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      *         no such value or it can't be converted to {@code float}.
      */
     @Nullable
-    public final Float getFloatAndRemove(CharSequence name) {
-        return contains(name) ? setters().getFloatAndRemove(name) : null;
-    }
+    Float getFloatAndRemove(CharSequence name);
 
     /**
      * Removes all the headers with the specified name and returns the header value which was added first.
@@ -469,9 +122,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @return the {@code float} value of the first value in insertion order or {@code defaultValue} if there
      *         is no such value or it can't be converted to {@code float}.
      */
-    public final float getFloatAndRemove(CharSequence name, float defaultValue) {
-        return contains(name) ? setters().getFloatAndRemove(name, defaultValue) : defaultValue;
-    }
+    float getFloatAndRemove(CharSequence name, float defaultValue);
 
     /**
      * Removes all the headers with the specified name and returns the header value which was added first.
@@ -481,9 +132,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      *         no such value or it can't be converted to {@code double}.
      */
     @Nullable
-    public final Double getDoubleAndRemove(CharSequence name) {
-        return contains(name) ? setters().getDoubleAndRemove(name) : null;
-    }
+    Double getDoubleAndRemove(CharSequence name);
 
     /**
      * Removes all the headers with the specified name and returns the header value which was added first.
@@ -493,9 +142,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @return the {@code double} value of the first value in insertion order or {@code defaultValue} if there
      *         is no such value or it can't be converted to {@code double}.
      */
-    public final double getDoubleAndRemove(CharSequence name, double defaultValue) {
-        return contains(name) ? setters().getDoubleAndRemove(name, defaultValue) : defaultValue;
-    }
+    double getDoubleAndRemove(CharSequence name, double defaultValue);
 
     /**
      * Removes all the headers with the specified name and returns the header value which was added first.
@@ -505,9 +152,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      *         value or it can't be converted to milliseconds.
      */
     @Nullable
-    public final Long getTimeMillisAndRemove(CharSequence name) {
-        return contains(name) ? setters().getTimeMillisAndRemove(name) : null;
-    }
+    Long getTimeMillisAndRemove(CharSequence name);
 
     /**
      * Removes all the headers with the specified name and returns the header value which was added first.
@@ -517,9 +162,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @return the milliseconds value of the first value in insertion order or {@code defaultValue} if there is
      *         no such value or it can't be converted to milliseconds.
      */
-    public final long getTimeMillisAndRemove(CharSequence name, long defaultValue) {
-        return contains(name) ? setters().getTimeMillisAndRemove(name, defaultValue) : defaultValue;
-    }
+    long getTimeMillisAndRemove(CharSequence name, long defaultValue);
 
     /**
      * Adds a new header with the specified {@code name} and {@code value}.
@@ -528,10 +171,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param value the header value
      * @return {@code this}
      */
-    public HttpHeadersBuilder add(CharSequence name, String value) {
-        setters().add(name, value);
-        return this;
-    }
+    HttpHeadersBuilder add(CharSequence name, String value);
 
     /**
      * Adds new headers with the specified {@code name} and {@code values}. This method is semantically
@@ -546,10 +186,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param values the header values
      * @return {@code this}
      */
-    public HttpHeadersBuilder add(CharSequence name, Iterable<String> values) {
-        setters().add(name, values);
-        return this;
-    }
+    HttpHeadersBuilder add(CharSequence name, Iterable<String> values);
 
     /**
      * Adds new headers with the specified {@code name} and {@code values}. This method is semantically
@@ -564,10 +201,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param values the header values
      * @return {@code this}
      */
-    public HttpHeadersBuilder add(CharSequence name, String... values) {
-        setters().add(name, values);
-        return this;
-    }
+    HttpHeadersBuilder add(CharSequence name, String... values);
 
     /**
      * Adds all header names and values of the specified {@code headers}.
@@ -575,11 +209,8 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @return {@code this}
      * @throws IllegalArgumentException if {@code headers == this}.
      */
-    public HttpHeadersBuilder add(
-            Iterable<? extends Entry<? extends CharSequence, String>> headers) {
-        setters().add(headers);
-        return this;
-    }
+    HttpHeadersBuilder add(
+            Iterable<? extends Entry<? extends CharSequence, String>> headers);
 
     /**
      * Adds a new header. Before the {@code value} is added, it's converted to a {@link String}.
@@ -588,10 +219,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param value the header value
      * @return {@code this}
      */
-    public HttpHeadersBuilder addObject(CharSequence name, Object value) {
-        setters().addObject(name, value);
-        return this;
-    }
+    HttpHeadersBuilder addObject(CharSequence name, Object value);
 
     /**
      * Adds a new header with the specified name and values. This method is equivalent to
@@ -605,10 +233,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param values the header values
      * @return {@code this}
      */
-    public HttpHeadersBuilder addObject(CharSequence name, Iterable<?> values) {
-        setters().addObject(name, values);
-        return this;
-    }
+    HttpHeadersBuilder addObject(CharSequence name, Iterable<?> values);
 
     /**
      * Adds a new header with the specified name and values. This method is equivalent to
@@ -622,10 +247,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param values the header values
      * @return {@code this}
      */
-    public HttpHeadersBuilder addObject(CharSequence name, Object... values) {
-        setters().addObject(name, values);
-        return this;
-    }
+    HttpHeadersBuilder addObject(CharSequence name, Object... values);
 
     /**
      * Adds all header names and values of the specified {@code headers}.
@@ -633,10 +255,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @return {@code this}
      * @throws IllegalArgumentException if {@code headers == this}.
      */
-    public HttpHeadersBuilder addObject(Iterable<? extends Entry<? extends CharSequence, ?>> headers) {
-        setters().addObject(headers);
-        return this;
-    }
+    HttpHeadersBuilder addObject(Iterable<? extends Entry<? extends CharSequence, ?>> headers);
 
     /**
      * Adds a new header.
@@ -645,10 +264,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param value the header value
      * @return {@code this}
      */
-    public HttpHeadersBuilder addInt(CharSequence name, int value) {
-        setters().addInt(name, value);
-        return this;
-    }
+    HttpHeadersBuilder addInt(CharSequence name, int value);
 
     /**
      * Adds a new header.
@@ -657,10 +273,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param value the header value
      * @return {@code this}
      */
-    public HttpHeadersBuilder addLong(CharSequence name, long value) {
-        setters().addLong(name, value);
-        return this;
-    }
+    HttpHeadersBuilder addLong(CharSequence name, long value);
 
     /**
      * Adds a new header.
@@ -669,10 +282,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param value the header value
      * @return {@code this}
      */
-    public HttpHeadersBuilder addFloat(CharSequence name, float value) {
-        setters().addFloat(name, value);
-        return this;
-    }
+    HttpHeadersBuilder addFloat(CharSequence name, float value);
 
     /**
      * Adds a new header.
@@ -681,10 +291,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param value the header value
      * @return {@code this}
      */
-    public HttpHeadersBuilder addDouble(CharSequence name, double value) {
-        setters().addDouble(name, value);
-        return this;
-    }
+    HttpHeadersBuilder addDouble(CharSequence name, double value);
 
     /**
      * Adds a new header.
@@ -693,10 +300,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param value the header value
      * @return {@code this}
      */
-    public HttpHeadersBuilder addTimeMillis(CharSequence name, long value) {
-        setters().addTimeMillis(name, value);
-        return this;
-    }
+    HttpHeadersBuilder addTimeMillis(CharSequence name, long value);
 
     /**
      * Sets a header with the specified name and value. Any existing headers with the same name are
@@ -706,10 +310,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param value the header value
      * @return {@code this}
      */
-    public HttpHeadersBuilder set(CharSequence name, String value) {
-        setters().set(name, value);
-        return this;
-    }
+    HttpHeadersBuilder set(CharSequence name, String value);
 
     /**
      * Sets a new header with the specified name and values. This method is equivalent to
@@ -724,10 +325,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param values the header values
      * @return {@code this}
      */
-    public HttpHeadersBuilder set(CharSequence name, Iterable<String> values) {
-        setters().set(name, values);
-        return this;
-    }
+    HttpHeadersBuilder set(CharSequence name, Iterable<String> values);
 
     /**
      * Sets a header with the specified name and values. Any existing headers with the specified name are
@@ -743,10 +341,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param values the header values
      * @return {@code this}
      */
-    public HttpHeadersBuilder set(CharSequence name, String... values) {
-        setters().set(name, values);
-        return this;
-    }
+    HttpHeadersBuilder set(CharSequence name, String... values);
 
     /**
      * Retains all current headers but calls {@link #set(CharSequence, String)} for each entry in
@@ -755,10 +350,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param headers the headers used to set the header values
      * @return {@code this}
      */
-    public HttpHeadersBuilder set(Iterable<? extends Entry<? extends CharSequence, String>> headers) {
-        setters().set(headers);
-        return this;
-    }
+    HttpHeadersBuilder set(Iterable<? extends Entry<? extends CharSequence, String>> headers);
 
     /**
      * Copies the entries missing in this headers from the specified headers.
@@ -773,10 +365,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      *
      * @return {@code this}
      */
-    public HttpHeadersBuilder setIfAbsent(Iterable<? extends Entry<? extends CharSequence, String>> headers) {
-        setters().setIfAbsent(headers);
-        return this;
-    }
+    HttpHeadersBuilder setIfAbsent(Iterable<? extends Entry<? extends CharSequence, String>> headers);
 
     /**
      * Sets a new header. Any existing headers with the specified name are removed. Before the {@code value} is
@@ -786,10 +375,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param value the value of the header
      * @return {@code this}
      */
-    public HttpHeadersBuilder setObject(CharSequence name, Object value) {
-        setters().setObject(name, value);
-        return this;
-    }
+    HttpHeadersBuilder setObject(CharSequence name, Object value);
 
     /**
      * Sets a header with the specified name and values. Any existing headers with the specified name are
@@ -805,10 +391,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param values the values of the header
      * @return {@code this}
      */
-    public HttpHeadersBuilder setObject(CharSequence name, Iterable<?> values) {
-        setters().setObject(name, values);
-        return this;
-    }
+    HttpHeadersBuilder setObject(CharSequence name, Iterable<?> values);
 
     /**
      * Sets a header with the specified name and values. Any existing headers with the specified name are
@@ -824,10 +407,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param values the values of the header
      * @return {@code this}
      */
-    public HttpHeadersBuilder setObject(CharSequence name, Object... values) {
-        setters().setObject(name, values);
-        return this;
-    }
+    HttpHeadersBuilder setObject(CharSequence name, Object... values);
 
     /**
      * Retains all current headers but calls {@link #setObject(CharSequence, Object)} for each entry in
@@ -836,10 +416,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param headers the headers used to set the values in this instance
      * @return {@code this}
      */
-    public HttpHeadersBuilder setObject(Iterable<? extends Entry<? extends CharSequence, ?>> headers) {
-        setters().setObject(headers);
-        return this;
-    }
+    HttpHeadersBuilder setObject(Iterable<? extends Entry<? extends CharSequence, ?>> headers);
 
     /**
      * Sets a header with the specified {@code name} to {@code value}. This will remove all previous values
@@ -849,10 +426,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param value the header value
      * @return {@code this}
      */
-    public HttpHeadersBuilder setInt(CharSequence name, int value) {
-        setters().setInt(name, value);
-        return this;
-    }
+    HttpHeadersBuilder setInt(CharSequence name, int value);
 
     /**
      * Sets a header with the specified {@code name} to {@code value}. This will remove all previous values
@@ -862,10 +436,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param value the header value
      * @return {@code this}
      */
-    public HttpHeadersBuilder setLong(CharSequence name, long value) {
-        setters().setLong(name, value);
-        return this;
-    }
+    HttpHeadersBuilder setLong(CharSequence name, long value);
 
     /**
      * Sets a header with the specified {@code name} to {@code value}. This will remove all previous values
@@ -875,10 +446,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param value the header value
      * @return {@code this}
      */
-    public HttpHeadersBuilder setFloat(CharSequence name, float value) {
-        setters().setFloat(name, value);
-        return this;
-    }
+    HttpHeadersBuilder setFloat(CharSequence name, float value);
 
     /**
      * Sets a header with the specified {@code name} to {@code value}. This will remove all previous values
@@ -888,10 +456,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param value the header value
      * @return {@code this}
      */
-    public HttpHeadersBuilder setDouble(CharSequence name, double value) {
-        setters().setDouble(name, value);
-        return this;
-    }
+    HttpHeadersBuilder setDouble(CharSequence name, double value);
 
     /**
      * Sets a header with the specified {@code name} to {@code value}. This will remove all previous values
@@ -901,10 +466,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param value the header value
      * @return {@code this}
      */
-    public HttpHeadersBuilder setTimeMillis(CharSequence name, long value) {
-        setters().setTimeMillis(name, value);
-        return this;
-    }
+    HttpHeadersBuilder setTimeMillis(CharSequence name, long value);
 
     /**
      * Removes all headers with the specified {@code name}.
@@ -912,9 +474,7 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param name the header name
      * @return {@code true} if at least one entry has been removed.
      */
-    public final boolean remove(CharSequence name) {
-        return contains(name) ? setters().remove(name) : false;
-    }
+    boolean remove(CharSequence name);
 
     /**
      * Removes all headers with the specified {@code name}. Unlike {@link #remove(CharSequence)}
@@ -923,27 +483,12 @@ public class HttpHeadersBuilder implements HttpHeaderGetters {
      * @param name the header name
      * @return {@code this}
      */
-    public HttpHeadersBuilder removeAndThen(CharSequence name) {
-        if (contains(name)) {
-            setters().remove(name);
-        }
-        return this;
-    }
+    HttpHeadersBuilder removeAndThen(CharSequence name);
 
     /**
      * Removes all headers. After a call to this method, {@link #size()} becomes {@code 0}.
      *
      * @return {@code this}
      */
-    public HttpHeadersBuilder clear() {
-        if (!isEmpty()) {
-            setters().clear();
-        }
-        return this;
-    }
-
-    @Override
-    public String toString() {
-        return getClass().getSimpleName() + getters();
-    }
+    HttpHeadersBuilder clear();
 }
